requirements.txt
sqlmodel
python-dotenv
pillow
requests
fastapi
uvicorn[standard]
pydantic>=2.0.0
